Bugfix: Das Abstellen von mehrere Modulen nun endlich Fehlerfrei

+++ htdocs/admin_modules.php
@@ -63,18 +63,45 @@ if ($perm->have_studip_perm("tutor", $admin_modules_data["range_id"])) {
 	if ($default_x) {
 		$admin_modules_data["changed_bin"] = $amodules->getDefaultBinValue($admin_modules_data["range_id"]);
 	}
+	
+	//consistency: kill objects
+	foreach ($amodules->registered_modules as $key => $val) {
+		$moduleXxDeactivate = "module".$key."Deactivate";
+		$delete_xx = "delete_".$key;
+	
+		if (($$delete_xx) && (method_exists($amodules,$moduleXxDeactivate))) {
+			$amodules->$moduleXxDeactivate($admin_modules_data["range_id"]);
+			$amodules->clearBit($admin_modules_data["changed_bin"], $amodules->registered_modules[$key]["id"]);
+			unset($admin_modules_data["conflicts"][$key]);
+			$resolve_comflicts = TRUE;
+		}
+	}
 
+	//consitency: cancel kill objects
+	foreach ($amodules->registered_modules as $key => $val) {
+		$cancel_xx = "cancel_".$key;
+		
+		if (($$cancel_xx) && (method_exists($amodules,$moduleXxDeactivate))) {
+			$amodules->setBit($admin_modules_data["changed_bin"], $amodules->registered_modules[$key]["id"]);
+			unset($admin_modules_data["conflicts"][$key]);
+			$resolve_comflicts = TRUE;
+		}
+	}
+	
 	if (($uebernehmen_x) || ($retry)) {
 		$msg='';
 
 		if ($uebernehmen_x)
 			foreach ($amodules->registered_modules as $key => $val) {
+				//after sending, set all "conflicts" to TRUE (we check them later)
+				$admin_modules_data["conflicts"][$key] = TRUE;
+				
 				$tmp_key = $key."_value";
 				if ($$tmp_key == "TRUE")
 					$$tmp_key = TRUE;
 				else
 					$$tmp_key = FALSE;
-		
+
 				if ($$tmp_key) {
 					$amodules->setBit($admin_modules_data["changed_bin"], $amodules->registered_modules[$key]["id"]);
 				} else {
@@ -88,21 +115,21 @@ if ($perm->have_studip_perm("tutor", $admin_modules_data["range_id"])) {
 			$cancel_xx = "cancel_".$key;
 			
 			//checks for deactivating a module
-			if ((!$$delete_xx) && (!$$cancel_xx)) {
-				$getModuleXxExistingItems = "getModule".$key."ExistingItems";
-		
-				if (method_exists($amodules,$getModuleXxExistingItems)) {
-					if (($amodules->isBit($admin_modules_data["orig_bin"],  $amodules->registered_modules[$key]["id"])) &&
-						(!$amodules->isBit($admin_modules_data["changed_bin"],  $amodules->registered_modules[$key]["id"])) &&
-						($amodules->$getModuleXxExistingItems($admin_modules_data["range_id"]))) {
-						
-						$msg.="info§".$amodules->registered_modules[$key]["msg_warning"];
-						$msg.="<br /><a href=\"".$PHP_SELF."?delete_$key=TRUE&retry=TRUE\">" . makeButton("ja2", "img") . "</a>&nbsp; \n";
-						$msg.="<a href=\"".$PHP_SELF."?cancel_$key=TRUE&retry=TRUE\">" . makeButton("nein", "img") . "</a>\n§";
-						$dont_save = TRUE;
-					}
-				}
-			}
+			$getModuleXxExistingItems = "getModule".$key."ExistingItems";
+	
+			if (method_exists($amodules,$getModuleXxExistingItems)) {
+				if (($amodules->isBit($admin_modules_data["orig_bin"],  $amodules->registered_modules[$key]["id"])) &&
+					(!$amodules->isBit($admin_modules_data["changed_bin"],  $amodules->registered_modules[$key]["id"])) &&
+					($amodules->$getModuleXxExistingItems($admin_modules_data["range_id"])) &&
+					($admin_modules_data["conflicts"][$key])) {
+					
+					$msg.="info§".$amodules->registered_modules[$key]["msg_warning"];
+					$msg.="<br /><a href=\"".$PHP_SELF."?delete_$key=TRUE&retry=TRUE\">" . makeButton("ja2", "img") . "</a>&nbsp; \n";
+					$msg.="<a href=\"".$PHP_SELF."?cancel_$key=TRUE&retry=TRUE\">" . makeButton("nein", "img") . "</a>\n§";
+				} else
+					unset($admin_modules_data["conflicts"][$key]);
+			} else
+				unset($admin_modules_data["conflicts"][$key]);
 				
 			//checks for activating a module
 			$moduleXxActivate = "module".$key."Activate";
@@ -116,29 +143,8 @@ if ($perm->have_studip_perm("tutor", $admin_modules_data["range_id"])) {
 			}
 		}
 	}
-
-	//consistency: kill objects
-	foreach ($amodules->registered_modules as $key => $val) {
-		$moduleXxDeactivate = "module".$key."Deactivate";
-		$delete_xx = "delete_".$key;
 	
-		if (($$delete_xx) && (method_exists($amodules,$moduleXxDeactivate))) {
-			$amodules->$moduleXxDeactivate($admin_modules_data["range_id"]);
-			$amodules->clearBit($admin_modules_data["changed_bin"], $amodules->registered_modules[$key]["id"]);
-		}
-	}
-
-
-	//consitency: cancel kill objects
-	foreach ($amodules->registered_modules as $key => $val) {
-		$cancel_xx = "cancel_".$key;
-	
-		if (($$cancel_xx) && (method_exists($amodules,$moduleXxDeactivate))) {
-			$amodules->setBit($admin_modules_data["changed_bin"], $amodules->registered_modules[$key]["id"]);
-		}
-	}
-	
-	if ((!$dont_save) && ($admin_modules_data["orig_bin"] != $admin_modules_data["changed_bin"])) {
+	if ((!count($admin_modules_data["conflicts"])) && ($admin_modules_data["orig_bin"] != $admin_modules_data["changed_bin"])) {
 		$amodules->writeBin($admin_modules_data["range_id"], $admin_modules_data["changed_bin"]);
 		$admin_modules_data["orig_bin"] = $admin_modules_data["changed_bin"];
 		$admin_modules_data["modules_list"] = $amodules->getLocalModules($admin_modules_data["range_id"]);
@@ -157,7 +163,7 @@ if ($SessSemName[1])
 	$range_id=$SessSemName[1]; 
 	
 //wenn wir frisch reinkommen, werden benoetigte Daten eingelesen
-if (($range_id) && (!$uebernehmen_x) && (!$delete_forum) && (!$delete_documents)) {
+if (($range_id) && (!$uebernehmen_x) && (!$delete_forum) && (!$delete_documents) && (!count($admin_modules_data["conflicts"]))) {
 	$admin_modules_data["modules_list"] = $amodules->getLocalModules($range_id);
 	$admin_modules_data["orig_bin"] = $amodules->getBin($range_id);
 	$admin_modules_data["changed_bin"] = $amodules->getBin($range_id);
